+++ app/[locale]/projects/[projectId]/members/[memberId]/edit/__tests__/page.spec.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { act, fireEvent, render, screen, waitFor } from '@/utils/test-utils';
 import { useParams, useRouter } from 'next/navigation';
 import { useToast } from '@/context/ToastContext';
+import logECS from '@/utils/clientLogger';
 
 import {
   useProjectContributorQuery,
@@ -9,14 +10,18 @@ import {
   useUpdateProjectContributorMutation,
   useRemoveProjectContributorMutation
 } from '@/generated/graphql';
+import { useProjectContributorData } from '@/hooks/projectContributorData';
 
 import { axe, toHaveNoViolations } from 'jest-axe';
 import ProjectsProjectMembersEdit from '../page';
 import { mockScrollIntoView, mockScrollTo } from "@/__mocks__/common";
 import mockProjectContributorData from '../__mocks__/mockProjectContributorData.json';
 import mockContributorRoles from '../__mocks__/mockcontributorRoles.json';
+import mockResponse from '../__mocks__/mockResponseFromMutation.json';
 
 expect.extend(toHaveNoViolations);
+
+jest.mock('@/utils/clientLogger', () => jest.fn());
 
 
 jest.mock('next/navigation', () => ({
@@ -24,23 +29,6 @@ jest.mock('next/navigation', () => ({
   useRouter: jest.fn()
 }))
 
-const mockResponse = {
-  updateProjectContributor: {
-    givenName: "Juliet",
-    surName: "Cousteau",
-    orcid: "0000-JACQ-0000-0000",
-    id: 1,
-    errors: {
-      email: null,
-      surName: null,
-      general: null,
-      givenName: null,
-      orcid: null,
-      affiliationId: null,
-      contributorRoleIds: null,
-    },
-  }
-}
 // Mock the useTemplateQuery hook
 jest.mock("@/generated/graphql", () => ({
   useProjectContributorQuery: jest.fn(),
@@ -49,24 +37,8 @@ jest.mock("@/generated/graphql", () => ({
   useRemoveProjectContributorMutation: jest.fn()
 }));
 
-
 jest.mock('@/hooks/projectContributorData', () => ({
-  useProjectContributorData: () => ({
-    projectContributorData: {
-      givenName: 'Test',
-      surName: 'User',
-      affiliationId: '',
-      email: 'test@example.com',
-      orcid: '',
-    },
-    selectedRoles: [],
-    checkboxRoles: [],
-    setCheckboxRoles: jest.fn(),
-    loading: false,
-    setProjectContributorData: jest.fn(),
-    setSelectedRoles: jest.fn(),
-    queryError: null
-  })
+  useProjectContributorData: jest.fn() // Make it a mock function
 }));
 
 const mockRouter = {
@@ -99,13 +71,92 @@ describe("ProjectsProjectMembersEdit", () => {
       data: mockContributorRoles,
       loading: false,
       error: null,
-    }
-    );
+    });
+
+    (useProjectContributorData as jest.Mock).mockReturnValue({
+      projectContributorData: {
+        givenName: 'Test',
+        surName: 'User',
+        affiliationId: 'test-affiliation',
+        email: 'test@example.com',
+        orcid: '0000-0000-0000-0000',
+      },
+      checkboxRoles: ['1', '2'],
+      setCheckboxRoles: jest.fn(),
+      loading: false,
+      setProjectContributorData: jest.fn(),
+      data: {
+        projectContributor: {
+          givenName: 'Test',
+          surName: 'User',
+          affiliation: { uri: 'test-affiliation' },
+          email: 'test@example.com',
+          orcid: '0000-0000-0000-0000',
+          contributorRoles: [
+            { id: '1', __typename: 'ContributorRole' },
+            { id: '2', __typename: 'ContributorRole' }
+          ]
+        }
+      },
+      queryError: null
+    });
   });
 
 
   afterEach(() => {
     jest.clearAllMocks();
+  });
+
+  it('should render loading state', async () => {
+    (useContributorRolesQuery as jest.Mock).mockReturnValue({
+      data: null,
+      loading: true,
+      error: null,
+    });
+
+    (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
+      { loading: false, error: undefined },
+    ]);
+
+    (useRemoveProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
+      { loading: false, error: undefined },
+    ]);
+
+    await act(async () => {
+      render(
+        <ProjectsProjectMembersEdit />
+      );
+    });
+
+    expect(screen.getByText('messaging.loading...')).toBeInTheDocument();
+  });
+
+  it('should render error state', async () => {
+    (useContributorRolesQuery as jest.Mock).mockReturnValue({
+      data: null,
+      loading: false,
+      error: true,
+    });
+
+    (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
+      { loading: false, error: undefined },
+    ]);
+
+    (useRemoveProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
+      { loading: false, error: undefined },
+    ]);
+
+    await act(async () => {
+      render(
+        <ProjectsProjectMembersEdit />
+      );
+    });
+
+    expect(screen.getByText('messaging.error')).toBeInTheDocument();
   });
 
   it("should render correct fields", async () => {
@@ -169,7 +220,7 @@ describe("ProjectsProjectMembersEdit", () => {
     });
   });
 
-  it("should handle form submission errors", async () => {
+  it("should handle field level errors from submitting form", async () => {
     (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
       jest.fn().mockResolvedValueOnce({ data: { updateProjectContributor: { errors: { general: 'Error updating member' } } } }),
       { loading: false, error: undefined },
@@ -190,6 +241,37 @@ describe("ProjectsProjectMembersEdit", () => {
     fireEvent.click(saveButton);
 
     expect(await screen.findByText('Error updating member')).toBeInTheDocument();
+  });
+
+  it("should handle update member request errors", async () => {
+    (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockRejectedValueOnce(new Error("Error removing member")),
+      { loading: false, error: undefined },
+    ]);
+    (useRemoveProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
+      { loading: false, error: undefined },
+    ]);
+
+    await act(async () => {
+      render(
+        <ProjectsProjectMembersEdit />
+      );
+    });
+
+    const saveButton = screen.getByRole('button', { name: /buttons.saveChanges/i });
+    fireEvent.click(saveButton);
+
+    await waitFor(() => {
+      expect(logECS).toHaveBeenCalledWith(
+        'error',
+        'updateProjectContributor',
+        expect.objectContaining({
+          error: expect.anything(),
+          url: { path: '/projects/1/members/1/edit' },
+        })
+      );
+    });
   });
 
   it("should handle remove member", async () => {
@@ -217,7 +299,7 @@ describe("ProjectsProjectMembersEdit", () => {
     });
   });
 
-  it("should handle remove member errors", async () => {
+  it("should handle field-level errors from remove member request", async () => {
     (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
       jest.fn().mockResolvedValueOnce({ data: mockResponse }),
       { loading: false, error: undefined },
@@ -239,12 +321,94 @@ describe("ProjectsProjectMembersEdit", () => {
     expect(await screen.findByText('Error removing member')).toBeInTheDocument();
   });
 
+  it("should handle remove member request errors", async () => {
+    (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: mockResponse }),
+      { loading: false, error: undefined },
+    ]);
+    (useRemoveProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockRejectedValueOnce(new Error("Error removing member")),
+      { loading: false, error: undefined },
+    ]);
+
+    await act(async () => {
+      render(
+        <ProjectsProjectMembersEdit />
+      );
+    });
+
+    const removeButton = screen.getByRole('button', { name: /Remove member from project/i });
+    fireEvent.click(removeButton);
+
+    await waitFor(() => {
+      expect(logECS).toHaveBeenCalledWith(
+        'error',
+        'removeProjectContributor',
+        expect.objectContaining({
+          error: expect.anything(),
+          url: { path: '/projects/1/members/1/edit' },
+        })
+      );
+    });
+  });
+
+  it("should handle checkbox change", async () => {
+    (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: mockResponse }),
+      { loading: false, error: undefined },
+    ]);
+
+    (useRemoveProjectContributorMutation as jest.Mock).mockReturnValue([
+      jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
+      { loading: false, error: undefined },
+    ]);
+
+    const setCheckboxRolesMock = jest.fn();
+
+    // Override the mock implementation for this specific test
+    (useProjectContributorData as jest.Mock).mockReturnValue({
+      projectContributorData: {
+        givenName: 'Test',
+        surName: 'User',
+        affiliationId: 'test-affiliation',
+        email: 'test@example.com',
+        orcid: '0000-0000-0000-0000',
+      },
+      checkboxRoles: ['1', '2'],
+      setCheckboxRoles: setCheckboxRolesMock,
+      loading: false,
+      setProjectContributorData: jest.fn(),
+      data: {
+        projectContributor: {
+          givenName: 'Test',
+          surName: 'User',
+          affiliation: { uri: 'test-affiliation' },
+          email: 'test@example.com',
+          orcid: '0000-0000-0000-0000',
+          contributorRoles: [
+            { id: '1', __typename: 'ContributorRole' },
+            { id: '2', __typename: 'ContributorRole' }
+          ]
+        }
+      },
+      queryError: null
+    });
+
+    render(<ProjectsProjectMembersEdit />); // Ensure the component is rendered first
+
+    const checkboxes = screen.getAllByRole('checkbox');
+    checkboxes.forEach(checkbox => {
+      fireEvent.click(checkbox);
+    });
+
+    expect(setCheckboxRolesMock).toHaveBeenCalled();
+  });
+
   it("should pass accessibility checks", async () => {
     (useUpdateProjectContributorMutation as jest.Mock).mockReturnValue([
       jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),
       { loading: false, error: undefined },
     ]);
-
 
     (useRemoveProjectContributorMutation as jest.Mock).mockReturnValue([
       jest.fn().mockResolvedValueOnce({ data: { key: 'value' } }),