+++ hooks/projectContributorData.tsx
@@ -1,9 +1,10 @@
 'use client'
 
 import { useEffect, useState } from 'react';
-import { useProjectContributorQuery, ContributorRole } from '@/generated/graphql';
+import { useProjectContributorQuery } from '@/generated/graphql';
 import { ProjectContributorFormInterface } from '@/app/types';
 
+// Custom hook to fetch project contributor data
 export const useProjectContributorData = (projectContributorId: number) => {
   const [projectContributorData, setProjectContributorData] = useState<ProjectContributorFormInterface>({
     givenName: '',
@@ -14,7 +15,6 @@ export const useProjectContributorData = (projectContributorId: number) => {
   });
 
   // Keep track of which roles are checked
-  const [selectedRoles, setSelectedRoles] = useState<ContributorRole[]>([]);
   const [checkboxRoles, setCheckboxRoles] = useState<string[]>([]);
 
   // Query for the specified project contributor
@@ -23,9 +23,6 @@ export const useProjectContributorData = (projectContributorId: number) => {
       projectContributorId: Number(projectContributorId)
     }
   })
-
-
-  console.log("***DATA", data);
 
   useEffect(() => {
     // Update state values from data results
@@ -42,9 +39,6 @@ export const useProjectContributorData = (projectContributorId: number) => {
         const cleanedRoles = data.projectContributor?.contributorRoles.filter(role => role !== null && role !== undefined);
         /* eslint-disable no-unused-vars, @typescript-eslint/no-unused-vars */
         const cleanedData = cleanedRoles.map(({ __typename, ...fields }) => fields);
-        setSelectedRoles((prevRoles) => {
-          return [...prevRoles, ...cleanedData];
-        });
         const selectedRoleLabels = cleanedData
           .map(role => {
             return role?.id?.toString() ?? '';
@@ -57,12 +51,10 @@ export const useProjectContributorData = (projectContributorId: number) => {
 
   return {
     projectContributorData,
-    selectedRoles,
     checkboxRoles,
     setCheckboxRoles,
     loading,
     setProjectContributorData,
-    setSelectedRoles,
     data,
     queryError
   };